[BE] 로그인, 회원가입 기능 구현

+++ Balgoorm_BackEnd/src/main/java/com/balgoorm/balgoorm_backend/user/controller/UserController.java
@@ -30,7 +30,7 @@ public class UserController {
     }
 
     //전체 회원을 조회하는 API
-    @GetMapping("/admin/all")
+    @GetMapping("/admin/all") // 회원 정보에 비밀번호도 다 포함되어서 비밀번호는 뺄지 고민 중
     public ResponseEntity<List<User>> getAllMembers() {
         List<User> users = userService.getAllMembers();
         return ResponseEntity.ok(users);
@@ -64,29 +64,11 @@ public class UserController {
         return new ResponseEntity<>("회원가입 성공!", HttpStatus.CREATED);
     }
 
-    @PostMapping("/login")
-    public ResponseEntity<String> login(@RequestBody UserLoginRequest userLoginRequest) {
-        if (userLoginRequest.getUserId().isEmpty()) {
-            return new ResponseEntity<>("아이디가 비어있습니다", HttpStatus.BAD_REQUEST);
-        }
-        if (userLoginRequest.getPassword().isEmpty()) {
-            return new ResponseEntity<>("비밀번호가 비어있습니다", HttpStatus.BAD_REQUEST);
-        }
-
-        User user = userService.login(userLoginRequest);
-
-        if (user == null) {
-            return new ResponseEntity<>("로그인 실패", HttpStatus.UNAUTHORIZED);
-        }
-
-        log.info("로그인 {}", userLoginRequest);
-        return new ResponseEntity<>("로그인 성공", HttpStatus.OK);
-    }
 
 
     //회원 정보 수정 (닉네임, 비밀번호)
-
-    @PostMapping("/updateUser/{userid}")
+//비밀번호 닉네임 따로 변경
+    @PostMapping("/updateUser/{userid}") // 비밀번호와 닉네임 변경을 따로 만들지 고민 중
     public ResponseEntity<String> updateUser(@PathVariable("userid") String userId, @RequestBody UserUpdateRequest request) {
         userService.updateUser(request, userId);
         return ResponseEntity.ok("회원 정보가 수정되었습니다.");
@@ -95,9 +77,14 @@ public class UserController {
     //회원 탈퇴
     @DeleteMapping("/deleteUser/{userid}")
     public ResponseEntity<String> deleteUser(@PathVariable("userid") String userId, @RequestParam String password) {
-        userService.deleteUser(userId, password);
-        return ResponseEntity.ok("회원이 탈퇴되셨습니다.");
+        if (userService.checkPassword(userId, password)) {
+            userService.deleteUser(userId, password);
+            return ResponseEntity.ok("회원이 탈퇴되셨습니다.");
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("비밀번호가 일치하지 않습니다.");
+        }
     }
+
     @GetMapping("/current-user")
     public ResponseEntity<Map<String, String>> getCurrentUser(@AuthenticationPrincipal UserDetails userDetails) {
         Map<String, String> response = new HashMap<>();
